Changed profile pic trigger to create signed url
I was having the trigger convert the uploaded image to a public file with the same url. However, these makes it hard to determine when to invalidate the cached image on the client because the url never changes. Invalidating the cache every time the client loads an image creates a bad user experience due to how slow it is.