Fixed getSignedUrlBug
getSignedUrl now seems to produce the same url given the same input and
the same file. This might have been caused by the admin sdk update in
commit 374a64ab0c3ba9b8b63527f30acf2400a1569497. This prevents the user
document from changing when a new profile pic is uploaded, and so any
client listening for changes won't see any. Now, the expiration date of
the url is semi-randomized to some date in the range year 2500-12500,
resulting in a url that is likely unique. This range could potentially
be broadened to increase the randomness as long as it doesn't exceed
firebase's max date limit.