First draft of submitVote
Renamed status to gameStatus to be more specific because created snipeStatus constants.
createGame no creates the gameID with generateUniqueString() instead of using the default id generated by firebase so it's consistent with other unique id's we create.
game, snipe, and snipePicture documents now contain fields with their id. This makes it so you don't have to keep track of the doc ref associated with the doc data e.g. from doc.getData().
Added FCM utility functions for sending a message to a user given their uid and creating payload for snipe message.
Updated package-lock.json
Deleted shared.js because submitVote takes care of ending a game and so I didn't want to keep a file around with logic that was outdated and not being used.
submitSnipe now creates snipePicture document starting with ref count 0. Ref count is incremented when each snipe is created in a transaction. Each snipe is created in a different transaction.
submitSnipe now just sends the vote notification to the target instead of everyone else in the game.
First draft of submitVote. Still need to add error handling, but it should still work without that but the errors received on the client might not be as useful e.g. internal errors.