recursiveDelete -> clearFirestoreData
Recursive delte implementation didn't work for when documents didn't
exist anymore but had subcollections with documents that still existed.
the @firebase/testing library has a method to clear the data in the
emulator, so that is now used and all tests are passing.
Changed some names from snakecase to camelcase