some invalid gameIDs don't break submitSnipe
If on of the gameIDs have ended by the time the function runs, it would
cause Promise.all() to return the value of the rejected transaction.
This is incorrect because the user might be in other games that are
still going on. So, it now uses the reflect function to ignore invalid
gameIDs. This means that the client will not receive errors for invalid
gameIDs e.g. games that don't exist. We could read the game documents before
the transaction to check this, but that would result in more reads
overall.