dead players can now vote in 2nd-round vote
If voting is tied, the snipe is a failure.
Game now has numPlayers field to make calculating majority easier. This
is the total number of players in the game, including ones that have
died. If we allow people to leave the game, we will need to update this
when they leave.